舵机调整

+++ TEMPLATE_PROJECT/CODES/DRIVER/SERVO_MOTOR.h
@@ -3,8 +3,8 @@
  * @Author       : FZU Liao
  * @Date         : 2022-01-03 14:34:30
  * @LastEditors  : Liao
- * @LastEditTime : 2022-01-03 15:33:19
- * @FilePath     : \DRIVER\SERVO_MOTOR.h
+ * @LastEditTime : 2022-01-06 20:02:43
+ * @FilePath     : \TEMPLATE_PROJECT\CODES\DRIVER\SERVO_MOTOR.h
  * Copyright 2022 FZU Liao, All Rights Reserved. 
  */
 
@@ -21,11 +21,11 @@ void SERVO_INIT_MOTOR();
 
 /**
  * @description: 设置伺服电机工作角度
- * @param Angle uint8 |Angle|<=150
+ * @param Angle uint8 |Angle|<=10
  * @return *
  * @example: 
  */
-void SERVO_SET_ANGLE(short Angle);
+void SERVO_SET_ANGLE(int Angle);
 
 /**
  * @description: 逆初始化伺服电机
